Prevent clobbering of an existing database.yml

+++ tasks/wheres-your-database-yml-dude.rake
@@ -3,9 +3,12 @@ file "config/database.yml.example" do |task|
 end
 
 file "config/database.yml" => "config/database.yml.example" do |task|
-  puts "Dude! I found #{task.prerequisites.first}, so I'll make a copy of it for you."
-  cp task.prerequisites.first, task.name
-  abort "Make sure it's cromulent to your setup, then rerun the last command."
+  # This task will execute when the example file is newer than the main file (not just when it doesn't exist).
+  unless File.exists?(task.name)
+    puts "Dude! I found #{task.prerequisites.first}, so I'll make a copy of it for you."
+    cp task.prerequisites.first, task.name
+    abort "Make sure it's cromulent to your setup, then rerun the last command."
+  end
 end
 
 task :environment => "config/database.yml"
